🐚 UPDATE :: User image 및 작성자 일치 확인 코드

+++ src/main/java/com/dsm/up/domain/post/service/PostService.java
@@ -10,8 +10,11 @@ import com.dsm.up.domain.post.domain.type.StateType;
 import com.dsm.up.domain.post.exception.PostNotFoundException;
 import com.dsm.up.domain.post.presentation.dto.request.PostRequest;
 import com.dsm.up.domain.post.presentation.dto.response.PostResponse;
+import com.dsm.up.domain.user.exception.UserNotMatchException;
+import com.dsm.up.domain.user.service.util.UserUtil;
 import com.dsm.up.global.aws.S3Util;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.dsm.up.domain.post.presentation.dto.response.PostListResponse;
@@ -26,6 +29,9 @@ public class PostService {
 
     private final PostRepository postRepository;
     private final S3Util s3Util;
+    @Value("${cloud.aws.s3.default-image}")
+    private String defaultImage;
+    private UserUtil userUtil;
 
     @Transactional
     public Long create(PostRequest request, MultipartFile file) {
@@ -45,6 +51,7 @@ public class PostService {
     @Transactional // todo 수정하는 사람과 작성자가 일치하는 지 확인하기
     public Long update(Long id, PostRequest request, MultipartFile file) {
         Post post = postRepository.findById(id).orElseThrow(() -> PostNotFoundException.EXCEPTION);
+        if(!post.getUser().getAccountId().equals(userUtil.getUserId())) throw UserNotMatchException.EXCEPTION;
 
         if (file != null) {
             if(post.getPath() != null) s3Util.delete(post.getPath());
@@ -56,6 +63,7 @@ public class PostService {
     @Transactional //todo 삭제 요청하는 유저와 작성한 유저가 일치하는지 확인해야함
     public void delete(Long id) {
         Post post = postRepository.findById(id).orElseThrow(() -> PostNotFoundException.EXCEPTION);
+        if(!post.getUser().getAccountId().equals(userUtil.getUserId())) throw UserNotMatchException.EXCEPTION;
 
         s3Util.delete(post.getPath());
         postRepository.delete(post);
@@ -65,8 +73,10 @@ public class PostService {
     public PostResponse getPostDetails(Long id) {
         Post post = postRepository.findById(id).orElseThrow(() -> PostNotFoundException.EXCEPTION);
 
+
         return PostResponse.builder()
                 .userNickname(post.getUser().getNickname())
+                .profileImage(post.getUser().getPath().equals(defaultImage) ? post.getUser().getPath() : s3Util.getUrl(post.getUser().getPath()))
                 .createDate(post.getCreateDate())
                 .title(post.getTitle())
                 .content(post.getContent())
